requirements.txt
click==7.1.2
python-dotenv==0.20.0
selenium==4.1.3
selenium_wire==4.6.3
